Updated Floating Point Library for GainSensor. -FP Library contains one function, printFloat.
GainSensorFP now implements FP to do it's voltage conditioning, as far as I know, it works.

LongTest, is a small test routine for the FP and GSConditioning integration tests.

ModuloTest.c tests the pgmspace routines to see how they work with arrays.

uint16toa was added, and also a 'float_4dp' function. Which prints out first 4 decimal pts of the passed float.

uint8toa.c was changed to fix some the powers array so that the largest element is 100, not 10,000.